+++ src/engine/Graphics/Filtering.ts
@@ -16,3 +16,14 @@ export enum ImageFiltering {
    */
   Blended = 'Blended'
 }
+
+/**
+ * Parse the image filtering attribute value, if it doesn't match returns null
+ */
+export function parseImageFiltering(val: string): ImageFiltering | null {
+  switch (val) {
+    case ImageFiltering.Pixel: return ImageFiltering.Pixel;
+    case ImageFiltering.Blended: return ImageFiltering.Blended;
+    default: return null;
+  }
+}